requirements.txt
streamlit>=1.33.0
requests>=2.31.0
beautifulsoup4>=4.12.0
trafilatura>=1.8.0
python-dotenv>=1.0.1
openai>=1.30.0
